Split AppleWin -> Utilities for generic code not called by other emulator components.
