+++ AppleWinExpress2019.vcxproj.filters
@@ -214,6 +214,9 @@
     <ClCompile Include="source\Core.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="source\Utilities.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="source\AppleWin.h">
@@ -511,6 +514,9 @@
       <Filter>Source Files\Windows</Filter>
     </ClInclude>
     <ClInclude Include="source\Core.h">
+      <Filter>Source Files</Filter>
+    </ClInclude>
+    <ClInclude Include="source\Utilities.h">
       <Filter>Source Files</Filter>
     </ClInclude>
   </ItemGroup>
