+++ source/Utilities.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "Disk.h"
+#include "Harddisk.h"
+
+
+void LoadConfiguration();
+bool DoDiskInsert(const UINT slot, const int nDrive, LPCSTR szFileName);
+bool DoHardDiskInsert(const int nDrive, LPCSTR szFileName);
+void InsertFloppyDisks(const UINT slot, LPSTR szImageName_drive[NUM_DRIVES], bool& bBoot);
+void InsertHardDisks(LPSTR szImageName_harddisk[NUM_HARDDISKS], bool& bBoot);
+void UnplugHardDiskControllerCard(void);
