+++ source/Windows/WinVideo.cpp
@@ -32,6 +32,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "Windows/WinFrame.h"
 #include "AppleWin.h"
 #include "Video.h"
+#include "Core.h"
 #include "CPU.h"
 #include "Joystick.h"
 #include "Frame.h"
